Add loading indicators and organized component logic
