+++ client/src/Result.js
@@ -1,0 +1,79 @@
+import Table from '@mui/material/Table';
+import TableBody from '@mui/material/TableBody';
+import TableCell from '@mui/material/TableCell';
+import TableHead from '@mui/material/TableHead';
+import TableRow from '@mui/material/TableRow';
+import Box from '@mui/material/Box';
+import Collapse from '@mui/material/Collapse';
+import IconButton from '@mui/material/IconButton';
+import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
+import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
+import { useState, useEffect } from 'react';
+
+const getResultPorts = async (name, uuid) => {
+    let result = await fetch('http://localhost:8000/' + name + '/ports');
+    let content = result.json();
+    return content;
+};
+
+const Result = (props) => {
+    const { domain, uuid } = props;
+
+    const [open, setOpen] = useState(false);
+    const [cache, setCache] = useState([]);
+    const [loading, setLoading] = useState(false);
+
+    let hasScanned = false;
+
+    useEffect(() => {
+        const load = async () => {
+            setLoading(true);
+            const ports = await getResultPorts(domain, uuid);
+            setCache(ports);
+            setLoading(false);
+        };
+
+        if (open && !hasScanned) {
+            load();
+            hasScanned = true;
+        }
+    }, [open]);
+
+    return (<>
+        <TableRow sx={{ '& > *': { borderBottom: 'unset' } }}>
+            <TableCell>
+                <IconButton size="small" onClick={() => setOpen(!open)}>
+                    {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
+                </IconButton>
+            </TableCell>
+            <TableCell component="th" scope="row">
+                {domain}
+            </TableCell>
+        </TableRow>
+        <TableRow>
+            <TableCell style={{ paddingBottom: 0, paddingTop: 0}} colSpan={6}>
+                <Collapse in={open} timout="auto" unmountOnExit>
+                    <Box sx={{ margin: 1}}>
+
+                        {loading ?
+                        <p>Loading...</p>
+                        :
+                        <Table size="small">
+                            <TableHead>
+                                <TableRow>
+                                    <TableCell>Open Ports</TableCell>
+                                </TableRow>
+                            </TableHead>
+                            <TableBody>
+                                {cache.map(port => <TableRow key={port.port}>{port.port}</TableRow>)}
+                            </TableBody>
+                        </Table>
+                        }
+                    </Box>
+                </Collapse>
+            </TableCell>
+        </TableRow>
+    </>);
+};
+
+export default Result;
